Remove warnings

+++ MailGame/MailGame/src/Ui/Construction/Construction.cpp
@@ -8,7 +8,7 @@
 std::map<EntityTag, Construction::Recipe> Construction::recipes = {
 	{ EntityTag::PostOffice, {
 		[](Game* g, sf::Vector2f pos, IsoRotation rot) {
-			sf::Vector2i tilesPos(round(pos.x), round(pos.y));
+			sf::Vector2i tilesPos((int)round(pos.x), (int)round(pos.y));
 			std::shared_ptr<Entity> e = EntityPresets::postOffice(g, sf::Vector2f(tilesPos), rot);
 			g->getGameMap()->setBuildingForTile(tilesPos.x, tilesPos.y, e);
 			return e;
